fix insert_records and continue frontend work

+++ src/model/planos.py
@@ -1,10 +1,8 @@
-from tokenize import Double
-
 class Planos:
     def __init__(self,
 
                  nome_plano: str = None,
-                 valor_plano: Double = None
+                 valor_plano: int = None
                  ):
 
         self.set_nome_plano(nome_plano)
@@ -13,13 +11,13 @@ class Planos:
     def set_nome_plano(self, nome_plano: str):
         self.nome_plano = nome_plano
 
-    def set_valor_plano(self, valor_plano: Double):
+    def set_valor_plano(self, valor_plano: int):
         self.valor_plano = valor_plano
 
     def get_nome_plano(self) -> str:
         return self.nome_plano
 
-    def get_valor_plano(self) -> Double:
+    def get_valor_plano(self) -> int:
         return self.valor_plano
 
     def to_string(self) -> str:
